Twitter OAuth 1.0 example
Replaced the current example in the docs with code that can be directly
copied and pasted into a file. This makes it easier for someone to get
started without them worrying about running a unit test or installing
Mocha or whatever. They can just copy/paste, fill in the consumer key,
consumer secret, access token and access token secret and immediately be
able to access the Twitter API (at least for a single user account, see
more about that here: https://dev.twitter.com/oauth/overview/single-user)

Added it to the examples directory.